Remove binder priority propagation workaround
recent changes to how cgroups are handled caused mediaserver to generate
a lot of permission denied diagnostics. It highlighted code in the
IMediaMetadataRetriever  interface that wored around a binder /
scheduling priority problem from 2009 ['this is a temporary workaround'].
Further investigation showed that even before the newest patch, the
workaround was ineffective, but wasn't logging the failures.

This removes the 13+ year old temporary workaround that passed scheduling
policies between client and service; we now depend on the binder to
handle that correctly.

This will also stop the verbose logcat messages that brought this to our
attention.

Bug: 276957640
Test: cts-tradefed CtsMediaMiscTestCases
Google: 2526480
Change-Id: I992374350ce87f2751dd1367fc40c5622e9b8536